Score function interface (#13)
In #10 , the computation of the score is spread between
`linmod.eval.mae_proportions` and `eval.py`. The function
`mae_proportions()` returns a data frame that `eval.py` then needs to
summarize, in order to get a final score.

Instead, the score should be a single function that takes in samples and
data and returns a single score. This way, you can add in multiple
scores very easily.

---------

Co-authored-by: Thanasi Bakis <abakis@uci.edu>