YAML-run pipeline, part 2 (data preprocessing) (#28)
_This resolves #26._

Until now, we've been missing the ability to easily configure how data
is preprocessed.

The `linmod.data` script now accepts a single commandline argument for
the path to a YAML file configuring its behavior. This is optional;
without it, the default behavior (seen in the dictionary
`linmod.data.DEFAULT_CONFIG`) will be used. The YAML file only needs to
define the keys it wants to modify from default; missing keys will be
populated with the default values.

An example is given in `present-day-forecasting/config.yaml`. As
described in the README, this is run as `python3 linmod.data
config.yaml`.

---------

Co-authored-by: afmagee42 <148902749+afmagee42@users.noreply.github.com>