Data synchronization (#33)
This PR makes it so that `present-day-forecasting/main.py` generates its
own dataset.

Previously our workflow was to first download the data, then run the
`present-day-forecasting/main.py`. This allowed for the downloaded data
to correspond to a time-horizon that was different from what was
specified in `present-day-forecasting/config.yaml`.

I refactored `linmod.data` such that the functionality of
`linmod.data__main__` is now in a function `linmod.data.main()` which
`present-day-forecasting/main.py` calls using the config it was
provided.