Prior selection (#41)
This PR has suffered some scope creep, but the primary purpose is to
check the priors on the models.

## Priors
Priors were out of sync with changes made to the model structure, thus
some examination is necessary. I don't think this is the end-all be-all
of prior selection, but I do think it is an improvement.

Importantly, the more drastic MCMC convergence problems appear to have
been banished. At default run settings, convergence isn't quite what we
might like for production, but I'm not seeing any truly worrisome ESS
values in the dozens (low values now seem more like 200), or PSRF values
near 2 (high values now are more like 1.02).

The file `exploration/priors.qmd` contains explorations of what the
current priors imply.

### BaselineModel
No changes.

### IndependentDivisionsModel
I have tightened the variance on the slope prior. The original
Normal(0,1 allowed for a variant starting at 1% to hit ~100% in 1 week.
The current Normal(0,0.25^2) prior forbids one-week takeovers, though
does allow one-month takeovers.

#### Future note
A Normal(0, 0.1^2) would forbid one-month takeovers, which seems
reasonable to me. However, this choice appears to significantly hamper
model performance even with large amounts of data. This could be
deserved (that is, the model is missing some other factor) or
undeserved. We can pursue this when we work more seriously on model
development.

### HierarchicalDivisionsModel

#### Intercepts
I have changed the intercept prior drastically in order to maintain some
semblance of parity with the non-hierarchical model. By "parity" I mean
"the marginal distributions on the $\beta_{0 \ell}$ terms are the same."

(On its own, this does not seem to make a huge difference in parameter
estimates or forecast scoring where tested.)

#### Slopes
I have changed both prior distributions here and streamlined the model,
removing unneeded free parameters.

Since all lineages are being treated equivalently, we do not need as
much hierarchical flexibility above `mu_beta_1`, which is now just
sampled directly from a Normal with fixed hyperparameters, rather than a
Normal with variable ones. This also reduces the variance on the slopes
towards more reasonable parameter values, and removes 2 parameters,
streamlining the DAG a bit.

Beyond this, I moved us to a 0-mode prior for the variance component of
the MVN prior on $\boldsymbol{\Sigma}$, where we had one which favored
quite large variances. The mean of this prior is set such that the
marginal variance on the slopes matches the independent model at the
mean. (The model pulls this up, which I am not pleased with, but with
which we can deal later.)

#### Future note
We may want to consider:
- Revisiting the prior on the diagonal of $\boldsymbol{\Sigma}$.
- Adaptive strength of pooling. It is currently a function of the priors
on pooled vs unpooled terms, but that need not be true.
- Letting the LKJ concentration parameter be inferred.

## Scope creep
The scope-creep comes from the need to better-understand what the models
"see" when fitting, since some values of hyperparameters that are, _a
priori_ quite reasonable seem potentially overly restrictive. So,
visualization tooling has been expanded so that:
1. We can plot our forecasts with data, this is split into:
1. Adding the option to plot data used to fit the model to
`forecast_{model_name}.png`
2. Adding an `eval_{model_name}.png` that plots the parameters with the
evaluation data for the forecast period.
2. The pure-data-viz script has been expanded to allow filtering to data
available during the period and only including specific locations
(though it remains a location-aggregated tool for now)

I have chosen to split 1 into 1.1 and 1.2 for now because I think it's
important to understand what was input data and what was not, and our
situation makes that otherwise hard (high-dimensional models with many
ways to slice things, plus there is significant temporal overlap between
our evaluation and modeling datasets).

---------

Co-authored-by: Thanasi Bakis <abakis@uci.edu>