Enable fitting to Omicron (#45)
This PR identifies two sources of difficulty in fitting the model in
early 2022 (end of Delta, start of Omicron).
1. Data filtering now allows removing trivial lineages and grouping them
into "other," resolving #29 and greatly reducing the computational
burden when many negligible lineages are floating around.
2. The hierarchical model appears to have been a bit too flexible, even
with the changes in #41. Here we remove one layer of the hierarchy,
fixing `sigma_beta_1` instead of inferring it.

The combined result is that the hierarchical model now works (MCMC is
believable) and produces reasonable-seeming results for a 2022-01-01
forecast date.

Late addition, mostly out of scope but worth including: This PR also
removes the filtering based on comparing sequence date to clade name.
The intent was to avoid clearly incorrect calls like 23A in April 2020.
However, it was causing problems for lineages like 24A, which takes off
in late 2023 and starts 2024 at high prevalence. As the percent of all
instances of (clade year) > (sample year) is small, and as many of those
instances are clearly valid, leaving the remainder in the dataset is the
lesser evil. Especially with the institution of (1) which should sweep
those into "other," minimizing issues.

---------

Co-authored-by: Thanasi Bakis <abakis@uci.edu>