End-to-end test (#81)
A test of the `retrospective-forecasting` pipeline running end to end
without failure.
Closes #43.

A YAML configuration and corresponding raw dataset is provided in
`retrospective-forecasting/test/`.
The test itself is added as a GitHub Actions workflow step.

The raw data provided to the pipeline was formed by filtering nextstrain
data dates to the interval [2021, 2023), filtering divisions to
{California, New York}, and randomly subsampling 10% of the resulting
rows.