Adding realism to retrospective data (#83)
This PR implements the two-step solution to the retrospective data
problem outlined in #75. In short, the problems are:
1. Clade assignments are nominally unstable for sequences older than a
year or so
2. More recent clade assignments don't capture the as-yet-unrecognized
taxon problem, where a soon-to-be-recognized taxon is growing but this
is hidden in its parent taxon

The solution implemented:
1. Allows us to obtain clade assignments from closer in time to the
`forecast_date` using UShER metadata.
2. Use [CDCgov/cladecombiner](https://github.com/CDCgov/cladecombiner)
to collapse taxa into only the ancestors which were recognized as of the
`forecast_date`

Beyond adding new functionalty to `linmod.data.main` to download from
UShER, and `DEFAULT_CONFIG` options to control the new bits, I
refactored a bit to pull into separate functions:
- Existing Nextstrain processing
- Clade assignment swapping with UShER
- Clade combining with cladecombiner

I think this adds some conceptual clarity to what's happening, and it
made it easier to check the effects with `exploration/whole_hist_viz.py`
which can now make use of all that directly.

I'm marking this as a draft for now since I think it's in our best
interests to get #81 in first. The new features are optional, so we can
still get the tests to pass.

### Possibly out of scope?

Working on this highlighted to me some potential brewing tensions on
dates, versioning, and vintaging in the pipeline. The downloading of
UShER data tracks its vintage per our [notation
doc](https://github.com/CDCgov/cfa-viral-lineage-model/blob/main/docs/notation.md#dates),
but we probably ought to be more careful about tracking Nextstrain's as
well, plus recording somewhere in the outputs of a pipeline run what was
used.

---------

Co-authored-by: Thanasi Bakis <abakis@uci.edu>