Re-enable selection of lineages (#91)
Fixes a bug introduced in #83 in which the recoding of lineages not in
`config["data"]["lineages"]` was ignored.

Cause: other-recoding was applied to the Nextstrain data, but labels are
then (potentially) taken from UShER and then (potentially) adjusted by
cladecombiner.

Solution: recode after all other `lineage` label modifications are
complete.