Small patches (#102)
This PR:
- Passes the seed to the multinomial sampler in `CountsEvaluator`,
resolving #100
- Ensures all evaluable divisions are in the modeling data by making
sure _all_ divisions (present in the evaluation data) have count data on
the `forecast_date`. As turnaround times are never that fast, in
practice this means 0 counts for all lineages in all divisions on that
date. But it is implemented as "if not present, add 0 counts" just in
case.
- Adds some paranoid checks and warnings, as suggested by the discussion
of #101. Now running the `taxonomy_shift` example prints
  ```
The following divisions have no data and have been dropped from the
evaluation set: {'Idaho', 'Maine', 'Hawaii', 'Wyoming'}
  ```
- Passes a seed to the forecast frame validation test, hopefully
resolving the unreproducible failure of pytest on main
- Note that I'm not sure _how_ a particular seed made pytest fail, since
the fake data was normalized. But this seems to have worked.