Scoring changes (#107)
This PR:
- Adds an absolute scoring metric, the proportion of uncovered
predictions, resolving #106
- Deprecates the proportions energy score and associated tests,
resolving #105
- Adds machinery to filter score aggregations by lineages, divisions,
and/or days, such that #104 can be resolved purely by changes to
`retrospective-forecasting.main` and the input yamls.
- Deprecates some `LazyFrame` type hints that were making type checking
in pylance very unhappy.
- Includes some random formatting changes `black` has gotten mad about
since we last touched this code.

Out of scope:
- In removing the proportion energy score test, it becomes clear that we
did not test the count-based energy score. We should remedy this. But
getting that right is also dependent on _how_ we handle the energy
score, which is currently in question.