Allow custom lineage names (#109)
This PR adds the ability for users to relabel lineages in a custom way.
This happens after UShER relabeling and before cladecombiner is called.

The order of operations here is important, since using UShER labels
overwrites lineage labels wholesale, and if we did this after
cladecombiner as-of aggregation, we could not resolve #103 with this
change.

If not using cladecombiner aggregation, this enables arbitary/custom
aggregation of lineages, should that be desired.