Change CI alpha for proportion uncovered evaluator (#110)
Preliminary results (e.g.
https://github.com/CDCgov/cfa-viral-lineage-model/pull/109#issuecomment-3169109087)
show that the forecasting distributions are overdispersed and cover more
than they should.

This is a bit hard to see with the previous 95% intervals, so this PR
swaps those for 89% intervals. E.g., the 95% interval for the Baseline
model in nomenclature-shift covers about 97%, which _could_ be fine
(it's only 2%, after all). but the 89% interval covers about 95% (and
that seems more clearly overcovered). Meanwhile the nominal 89% interval
for the ID model covers about 97%, which makes the point crystal clear.