Evaluating multiple different time ranges (#108)
This PR adjusts the pipeline, using the machinery from #107, so that we
produce 2 evaluation results, one for all division-days (prior to this
PR, this was what we were evaluating, overall model fit) and one for all
divisions but only on days with `fd_offset >= 1` (the forecasting
period, where there are no days which previously had any data).

It seemed easier to make the interface slightly broader, rather than
filter only on days to exactly specific ranges. So what is actually
implemented is the ability to pass filters for any column of the
combined forecast+count dataframe defining polars `.is_in()` filters.
Thus it could also be used to evaluate only certain lineages on certain
days in certain divisions.

Since #107 implements filtering when the statistic is computed, rather
than upon initializing the evaluation object, we only draw one set of
count data. This is both faster than drawing multiple and keeps the
statistics aligned, as they're based on all the same counts.

Resolves #104.